Complete rework
This is a complete rework of the synthesizer. It now operates on a 'node' based system for generating sound. A file format is also in development, but something for writing those files has not. 'Melody1.prm' was written in a hex editor, but is a valid song. A partial gui framework has also been developed to make it faster to add gui elements.