Keyboard GUI and change in library location
There is now an onscreen piano to accompany the keyboard, and the file structure has been changed to make it easier for others to build this project. SDL libraries are now directly in the project directory, though they will be empty on github to prevent redistribution. The window now also supports resizing.